tweaks

+++ src/assets/schemas/detectionNetPacket.capnp
@@ -7,7 +7,7 @@ struct DataPacket {
 }
 
 struct Detection {
-    id @0 :Int8;                          # Deep sort local id
+    id @0 :Int16;                          # Deep sort local id
     coordinates @1 :XYZ;                  # Tuple of integers for detection coordinates
     confidence @2 :Float64;                      # detection confidence
     isRobot @3 :Bool;                          # is robot detection flag
